requirements.txt
chromadb>=0.4.22
sentence-transformers>=2.2.2
openai>=1.12.0
python-dotenv>=1.0.0
pypdf>=4.0.0
llama-cpp-python>=0.3.0
huggingface_hub>=0.20.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9
streamlit>=1.28.0
